cleanup,org: remove inactive kubevirt org members
We have a policy for KubeVirt organization membership that requires
all members to be active contributors [1] .

In order to keep the community clean, inactive members will be removed
from time to time. That's what this commit does.

If you have contributions counting against the inactivity period, please
veto in this PR within a week.

[1]: https://github.com/kubevirt/community/blob/main/membership_policy.md#inactive-members

Signed-off-by: Daniel Hiller <dhiller@redhat.com>