flake-stats,ci: generate calendar page for flake stats
When we create issues for flaky tests for other teams to track and work
on we need to provide stable URLs to all reports.

This contradicts with the current entry html page we use to browse
towards todays flake-stats report [1] (which is unstable anyway since
the file is overwritten every day). Therefore we manually replace the
URL with that of the one containing the date of the report [2].

This commit replaces the double copy of the flake-stats report for
today with a generated html page that contains a calendar with links to
all reports, in descending timely order - that calendar page is then
placed into the entry URL [1].

[1]: https://storage.googleapis.com/kubevirt-prow/reports/flakefinder/kubevirt/kubevirt/flake-stats-14days.html
[2]: https://storage.googleapis.com/kubevirt-prow/reports/flakefinder/kubevirt/kubevirt/flake-stats-14days-2024-11-11.html

Signed-off-by: Daniel Hiller <dhiller@redhat.com>