needs-approver-review: avoid labelling PRs with changes requested
GitHub search currently lacks being able to query for PRs that are
missing a review by an approver, since the concept of approver vs.
reviewer is introduced via OWNERS, which GitHub doesn't know about.

This change adds a condition to the search query that selects PRs to
label with `needs-approver-review` in order to exclude PRs where changes
have been requested.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>