needs-approver-review: enable management by GH team
Currently the only user who can apply or remove the label
`needs-approver-review` is the `kubevirt-bot` user.

This change introduces a new GitHub team under the KubeVirt
org that contains the people who will be allowed to apply or
remove the label - changing the orgs.yaml to achieve this.

Then the `label` config is extended to allow above created group to
manage the label - changing plugins.yaml.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>