skip-review,prow,bumps: create initial configuration (#3789)
This commit introduces a label configuration for a label
`skip-review` that is targeted at pull requests created
by automation doing simple bumping stuff.

It creates the label configuration for project-infra. It then adds a
configuration section for tide handling pull requests with that label.
It also adds a configuration to the label plugin that allows the
kubevirt-bot user to apply that label to pull requests.

Finally two jobs are modified to apply that label on creation of the
pull request.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>