Fix repeated jwt token generation
As we are calling get_access_token() almost in every function in gcs_restapi.sh, we are trying to get too many jwt tokens and there are limits to number of refresh tokens that are issued per client-user combination, and beyond limits it may make old token to stop working, we are re-using refresh token with this fix. Ref: https://developers.google.com/identity/protocols/oauth2#5.-refresh-the-access-token,-if-necessary.

Signed-off-by: cmerla <Chandra.Merla@redhat.com>