chore,image updates: fetch names from manifests (#3895)
* chore,deployment updates: fetch image names from manifests

Currently we rely on a fixed image list, which has to be modified
whenever a new image is added.

This change greps over the manifests and extracts all image names that
match quay.io/kubevirtci - this list is then used as input for the
update script.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* chore,image update: update prow jobs from manifests

We apply a similar change to the prow-jobs update script. Also we create
a script for a pr description containing as detail the images that were
updated in a list.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>