bootstrap: Increase max wait time for podman socket to be ready
We still occasionally see e2e lanes fail due to the podman socket not
being ready in time.
https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/13537/pull-kubevirt-e2e-k8s-1.31-sig-compute/1880903900127563776
https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/11266/pull-kubevirt-e2e-k8s-1.30-sig-storage/1881735319955443712

This is indicative of poor disk performance and high IO wait times.

In the short term, increase the wait timeout to stop failing the e2e
lanes failing.

In the long term, we need to investigate improving the storage setup on
the CI cluster

Signed-off-by: Brian Carey <bcarey@redhat.com>