bug,ginkgo: fix handling no ginkgo import (#3919)
After the changes around using the ginkgo command instead of the osexec
for ginkgo binary handling the case of ginkgo outline failing due to
missing import did not work any more. This commit fixes the issue by
recovering from the panic that is caused in the case, also returning the
error in all other cases by setting the named return argument.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>