cannier: map commits to test names
Adds the `extract changed-tests` command which determines for a given
revision range the changes to tests inside a code base.

Output is the names of the changed tests as well as the ginkgo node data
for the paths to each changed test. Paths are expanded from the ginkgo
tree, which means each path describes exactly one test.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>