skip-review: enable for k/cdi uploader prs (#3938)
We repeatedly had the issue for k/kubevirt, that binaries mirrored to the
gcs to avoid losing them were not reachable any more, as the mirroring PR
was not merged ahead of branching the release.

This change enables skip-review label, which will make tide merge PRs that
have all lanes succeeded and no forbidden labels, i.e.
`do-not-merge/hold`.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>