prow-workloads: Reduce the number of jobs that can be scheduled to each node (#4011)
Memory per node was doubled when the new prow workloads was created.
As we split up node resources by memory requests, this means that a
larger number of jobs are scheduled to each node. We have seen that this
increased workload has caused stability issues with the in cluster
container image cache[1][2] and has also significantly increased lane runtimes
when the cluster is loaded.

Increase the standard memory request from 29Gi to 52Gi to reduce the
number of jobs that can be scheduled to each node.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/14128/pull-kubevirt-e2e-k8s-1.32-sig-compute-serial/1897341187082162176
[2] https://search.ci.kubevirt.io/?search=net%2Fhttp%3A+TLS+handshake+timeout&maxAge=168h&context=1&type=build-log&name=&excludeName=&maxMatches=5&maxBytes=20971520&groupBy=job

Signed-off-by: Brian Carey <bcarey@redhat.com>