fix: add links to own prow and own merge workflow
We are currently relying on the Kubernetes managed prow and
documentation, but we want to link to our own prow instance and
documentation. This change links to the merge automation doc which
contains a description of the basic merge workflow.

It also links to our own prow command help page, since the installed
plugins differ from the main k8s prow instance.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>