prow,workloads: Increase resource requests/limits for image proxy and bazel cache (#4055)
These services run on a dedicated cache node in the workloads cluster -
it doesn't make sense to limit there resources so much.

It has also been seen that there are performance issues with the docker
image cache when a large number of jobs are run and the jobs begin to
fail to pull the images. When these issues occur, the
docker-mirror-proxy pod hits it's CPU limits and starts to throttle the
CPU usage.

Increase the resource requests & limits for both the docker-mirror-proxy
and greenhouse (bazel cache).

Signed-off-by: Brian Carey <bcarey@redhat.com>