fix(sig-compute,periodics): bump timeout values
As the sig-compute are on the brink of the timeout since a while now,
let's increase the lane timeout value to give the jobs a bit nore time
to finish.

We know that a failing test will likely increase the lane runtime by
around 5 minutes, so it makes sense to give the failing lanes a bit more
of headroom to finish writing their artifacts.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>