merge-commit-summary: initial prototype (#4185)
* feat(merge-commit-summary): create page generator

Creates a cli that from the set of merge commits of the latest merged
pull requests it scrapes the failed builds and displays them on one
page. Included are links to the build itself, to the pull request
history, and the artifacts of the build.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(merge-commit-summary): create periodic upload

We create a periodic prow job that generates the merge-commit-summary
page and uploads it to the gcs bucket.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(merge-commit-summary): add gh install

As the gh cli is not installed on our golang image we install it over
the rpm repository provided.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>