fix(arm,periodic): missing cgroup host mount
We are experiencing a lot less failures on the presubmit lane than on
the periodic. A comparison of the lane configurations yielded a
difference in the volume configurations - apparently the /sys/fs/cgroup
hostpath mount was missing. This commit adds it.

Also the audit log mount is changed from empty dir to hostpath mount.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>