create most-flaky-tests daily report (#4249)
* chore: refactor flake-stats into package

we want to reuse the types and the report generation, therefore we are
moving this into a package in order to be able to call it.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(ginkgo): add extraction code for test specs

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(quarantine-report): add most flaky tests report

Adds a cli tool that generates a report for the most flaky tests over
the last 14 days (by default).

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(most-flaky-tests,publish): add report generation

Add the job generating the daily report. Also add a link in the
reports.md doc.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>