feat(per-test-execution): determine k8s version for report (#4274)
* fix(body.Close): put into defer

By accident the body is closed before even an attempt to read it has
been made. Using a defer to close the body on returning from the func.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(auto-load-k8s): determine usable k8s version

As we need the sig-compute-migrations lane present we determine from the
latest stable the version that has it, and use that one as base for the
reports.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(per-test-execution): remove fixed k8s version

Since we determine the "good" k8s version from the latest stable and
then search for a lane of sig-compute-migrations, we don't need to
hard-code it any more.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>