disable PSA in release-1.1 job configurations (#4319)
there is a problem with pod bring up latency in release-1.1
the root cause comes from the VM based providers and the trigger
is usage of custom seccomp provider in the kubevirt e2e tests.

In order to keep the lanes stable we temporarily opt out from PSA
and from using seccomp.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>