disable PSA in release-1.0 job configurations
there is a problem with pod bring up latency in release-1.0
the root cause comes from the VM based providers and the trigger
is usage of custom seccomp provider in the kubevirt e2e tests.

In order to keep the lanes stable we temporarily opt out from PSA
and from using seccomp.

ref: https://github.com/kubevirt/project-infra/pull/4319

Signed-off-by: fossedihelm <ffossemo@redhat.com>