fix(cannier): treat rename correctly
If a set of commits contains a file rename that case needs to be
treated.

See https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/15478/pull-kubevirt-check-tests-for-flakes/1957829008544501760
for an example of a resulting error.

This commit basically adds handling of the porcelain parsing of the
file changes - case 'R'.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>