chore(slack reporting): flip reporting in prow config
Flips the logic of reporting to the kubevirt-ci-notifications CNCF
channel for kubevirt/kubevirt.

Instead of explicitly opting out of the reporting, we want to
only let periodics report that want to get reported. Thus we create a
kubevirt/kubevirt config, removing the periodics from the
job_types_to_report slice.

Also it removes the custom configs for the two specific repos and aligns
these into the global config.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>