chore(e2e-failures): split by main and release (#4390)
The core problem on the e2e job failures graph is that it’s showing
all branches - it should split failures on lanes between main and
release branches.

Reason is that not all stabilization PRs and flaky test fixes are
backported, so release branches are being affected more by flakes.

This change adds a couple of panels showing the numbers for the main
vs. the release branches.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>