feat(arm): re-add arm to the flake-stats
We had removed arm lane due to very high failure rates. Now that we are
continuously working on stabilizing the tests for arm and also
increasing the tests that are being run on arm, we should re-add that
lane to the list of lanes being reported over.

This change removes the filtering of the arm lane from the flake-stats
report generation periodic job.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>