feat(rehearse): collapse instructions
Currently whenever the rehearse plugin comments on the PR the
instructions section is added, however that section is quite big and
that bloats the comment.

This change puts the instructions on how to trigger rehearsals into a
GitHub style collapsed details section, so the comment is less chatty.

See https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/organizing-information-with-collapsed-sections
for details

Signed-off-by: Daniel Hiller <dhiller@redhat.com>