fix(fossa,goveralls): increase job resources (#4426)
We are seeing frequent failures in the push-kubevirt-fossa and
push-kubevirt-goveralls jobs. Noting that these are still running on the
control-plane cluster, they are likely failing due to resouce
constraints.

A CI search [1] shows that also the pull jobs are failing, not in the
frequence as the postsubmits, but still we think increasing memory
will stabilize them.

[1]: https://search.ci.kubevirt.io/?search=Process+did+not+finish+before+1h0m0s+timeout&maxAge=336h&context=1&type=build-log&name=%28push%7Cpull%29-kubevirt-%28fossa%7Cgoveralls%29&excludeName=&maxMatches=5&maxBytes=20971520&groupBy=job

Signed-off-by: Daniel Hiller <dhiller@redhat.com>