feat(arm): re-add arm to the flake reporting (#4397)
* feat(arm): re-add arm to the flake-stats

We had removed arm lane due to very high failure rates. Now that we are
continuously working on stabilizing the tests for arm and also
increasing the tests that are being run on arm, we should re-add that
lane to the list of lanes being reported over.

This change removes the filtering of the arm lane from the flake-stats
report generation periodic job.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix(arm): convert hard-coded lane filtering into flag

Instead of having the filtering buried inside the reporting code
hard-coded, we want to be able to configure it without touching the
code.

This change moves the configuration value into a constant, whereas it's
value is exposed as a cli flag.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix(arm): remove arm from lane filter default value

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>