mount local project-infra directory for building images (#3869)
* feat(publish_image.sh): mount project-infra directory

Instead of always checking out the main project-infra
repository just use the locally mounted directory if available.

This will enable to check the image build with the locally
provided source, which in turn will enable a presubmit job for the
image builds built from project-infra.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(publish_image.sh): add golang build env vars

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(prow,config): add job preset to add go env vars

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(Containerfile): align to using local project-infra

Changes the Containerfile files to use the project-infra if it is
available locally.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* feat(publish): create missing presubmit jobs

Create build jobs for images - now possible due to project-infra mounted
locally. Also update publish jobs to honor the new preset.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>