feat(label-checker): add -M flag for checking PR (#4589)
Adds the flag `-M` with an argument of a set of comma-separated
labels. If there exists a PR and has one of those, the PR creation will
be aborted.

This integrates the often-found conjunction of labels-checker calls with
git-pr.sh calls into the git-pr.sh script in order to avoid the
boilerplate of mentioning repo, branch name etc. twice.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>