fix(dequarantine,check-tests-for-flakes): timeout
As can be seen here [1] the internal Ginkgo timeout prohibits the lanes
running as long as the job timeout permits. This comes into play now
because INVOKE_FUNCTEST_ONCE only runs the process once, so the timeout
value is not valid per single test lane run but for all repetitions.

This change sets the env var KUBEVIRT_FUNC_TEST_GINKGO_TIMEOUT to a
value near the job timeout (-15 minutes less).

[1]: https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/16368/pull-kubevirt-check-dequarantine-test/2001230254634438656#1:build-log.txt%3A4176

Signed-off-by: Daniel Hiller <dhiller@redhat.com>