chore(grafana): improve e2e job failures/h
ATM the e2e job failures per hour pane (top on [1]) is showing the lanes
where failures happened, additionally it shows a sum of all failures/h.
This makes the failures on the lanes very had to distinguish. Same goes
for job abortions panel below.

This change updates the dashboard graphs in three ways - it removes the
sum graph, changes the display to stacking mode, which improves
overall usability of the graphs by showing each lane better, and finally
it adds a legend with last value property sorted descending.

[1]: https://grafana.ci.kubevirt.io/d/_6XBEfyIk/e2e-job-failures-h?orgId=1&from=now-6h&to=now&refresh=15m

Signed-off-by: Daniel Hiller <dhiller@redhat.com>